src/test/kotlin/com/clouddjr/advent2015/Day07Test.kt
package com.clouddjr.advent2015

import com.clouddjr.advent2015.Resources.resourceAsList
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

@DisplayName("Day 7")
class Day07Test {

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches actual`() {
            val answer = Day07(resourceAsList("day07.txt")).solvePart1()

            assertEquals(16_076, answer)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches actual`() {
            val answer = Day07(resourceAsList("day07.txt")).solvePart2()

            assertEquals(2_797, answer)
        }
    }
}